該当エンコード後文字のポジションがない場合を想定して-1を返すためにuintをintに変更
主に == で終わるエンコード文字がある時に、=を-1で判定して処理したいため

+++ esab46.go
@@ -149,9 +149,9 @@ func getChar(position uint) byte {
 	return encodeCharList[position]
 }
 
-func getPosition(charByte byte) uint {
+func getPosition(charByte byte) int {
 	encodeCharList := getEncodeCharList()
-	return uint(strings.IndexByte(encodeCharList, charByte))
+	return strings.IndexByte(encodeCharList, charByte)
 }
 
 func getEncodeCharList() string {
